BackendSrv: only add content-type on POST, PUT requests Fixes #22869
