+++ public/app/core/services/backend_srv.ts
@@ -1,4 +1,3 @@
-import omitBy from 'lodash/omitBy';
 import { from, merge, MonoTypeOperatorFunction, Observable, of, Subject, throwError } from 'rxjs';
 import { catchError, filter, map, mergeMap, retryWhen, share, takeUntil, tap, throwIfEmpty } from 'rxjs/operators';
 import { fromFetch } from 'rxjs/fetch';
@@ -14,6 +13,7 @@ import { ContextSrv, contextSrv } from './context_srv';
 import { coreModule } from 'app/core/core_module';
 import { Emitter } from '../utils/emitter';
 import { DataSourceResponse } from '../../types/events';
+import { parseInitFromOptions, parseUrlFromOptions } from '../utils/fetch';
 
 export interface DatasourceRequestOptions {
   retry?: number;
@@ -52,18 +52,6 @@ interface ErrorResponse<T extends ErrorResponseProps = any> {
 enum CancellationType {
   request,
   dataSourceRequest,
-}
-
-function serializeParams(data: Record<string, any>): string {
-  return Object.keys(data)
-    .map(key => {
-      const value = data[key];
-      if (Array.isArray(value)) {
-        return value.map(arrayValue => `${encodeURIComponent(key)}=${encodeURIComponent(arrayValue)}`).join('&');
-      }
-      return `${encodeURIComponent(key)}=${encodeURIComponent(value)}`;
-    })
-    .join('&');
 }
 
 export interface BackendSrvDependencies {
@@ -580,62 +568,3 @@ coreModule.factory('backendSrv', () => backendSrv);
 // Used for testing and things that really need BackendSrv
 export const backendSrv = new BackendSrv();
 export const getBackendSrv = (): BackendSrv => backendSrv;
-
-export const parseUrlFromOptions = (options: BackendSrvRequest): string => {
-  const cleanParams = omitBy(options.params, v => v === undefined || (v && v.length === 0));
-  const serializedParams = serializeParams(cleanParams);
-  return options.params && serializedParams.length ? `${options.url}?${serializedParams}` : options.url;
-};
-
-export const parseInitFromOptions = (options: BackendSrvRequest): RequestInit => {
-  const method = options.method;
-  const headers = parseHeaders(options);
-  const isAppJson = isContentTypeApplicationJson(headers);
-  const body = parseBody(options, isAppJson);
-
-  return {
-    method,
-    headers,
-    body,
-  };
-};
-
-export const parseHeaders = (options: BackendSrvRequest) => {
-  const headers = new Headers({
-    'Content-Type': 'application/json',
-    Accept: 'application/json, text/plain, */*',
-  });
-
-  if (options && options.headers) {
-    Object.keys(options.headers).forEach(key => {
-      headers.set(key, options.headers[key]);
-    });
-  }
-
-  return headers;
-};
-
-export const isContentTypeApplicationJson = (headers: Headers) => {
-  if (!headers) {
-    return false;
-  }
-
-  const contentType = headers.get('content-type');
-  if (contentType && contentType.toLowerCase() === 'application/json') {
-    return true;
-  }
-
-  return false;
-};
-
-export const parseBody = (options: BackendSrvRequest, isAppJson: boolean) => {
-  if (!options) {
-    return options;
-  }
-
-  if (!options.data || typeof options.data === 'string') {
-    return options.data;
-  }
-
-  return isAppJson ? JSON.stringify(options.data) : new URLSearchParams(options.data);
-};
